Use UserSettingsStore instead of UserSettingsStorage where possible
